# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Learning Management System (LMS) API
  description: |
    API for managing a Learning Management System with courses, modules, lessons, and user tracking.
    This API provides comprehensive functionality for creating and managing educational content,
    user enrollments, and progress tracking.
  version: 1.0.0
  contact:
    name: LMS API Support
    email: support@lmsapi.example.com
servers:
  - url: https://api.lms.example.com/v1
    description: Production server
  - url: https://staging-api.lms.example.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    Error:
      type: object
      properties:
        id:
          type: string
          example: api.course.get
        ver:
          type: string
          example: "1.0"
        ts:
          type: string
          format: date-time
        params:
          type: object
          properties:
            resmsgid:
              type: string
              format: uuid
            status:
              type: string
              enum: [failed]
            err:
              type: string
              example: RESOURCE_NOT_FOUND
            errmsg:
              type: string
              example: Course not found
        responseCode:
          type: integer
          example: 404
        result:
          type: object
    Course:
      type: object
      properties:
        courseId:
          type: string
          format: uuid
        tenentId:
          type: string
          format: uuid
        title:
          type: string
        alias:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        image:
          type: string
          nullable: true
        featured:
          type: boolean
          default: false
        free:
          type: boolean
          default: true
        certificateTerm:
          type: string
          enum: [NA, PASS_ALL, COMPLETE_ALL]
        certificateId:
          type: string
          nullable: true
        startDatetime:
          type: string
          format: date-time
        endDatetime:
          type: string
          format: date-time
        adminApproval:
          type: boolean
          default: false
        autoEnroll:
          type: boolean
          default: false
        status:
          type: string
          enum: [draft, published, unpublished, archived]
        params:
          type: object
          additionalProperties: true
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - title
        - shortDescription
        - description
        - startDatetime
        - endDatetime
    CourseList:
      type: object
      properties:
        count:
          type: integer
        courses:
          type: array
          items:
            type: object
            properties:
              courseId:
                type: string
                format: uuid
              title:
                type: string
              shortDescription:
                type: string
              image:
                type: string
                nullable: true
              featured:
                type: boolean
              free:
                type: boolean
              status:
                type: string
              startDatetime:
                type: string
                format: date-time
              endDatetime:
                type: string
                format: date-time
...
    SuccessResponse:
      type: object
      properties:
        id:
          type: string
        ver:
          type: string
        ts:
          type: string
          format: date-time
        params:
          type: object
          properties:
            resmsgid:
              type: string
              format: uuid
            status:
              type: string
              enum: [successful]
            err:
              type: object
            errmsg:
              type: string
        responseCode:
          type: integer
        result:
          type: object
...
security:
  - bearerAuth: []
paths:
  /courses:
    get:
      summary: Get all courses
      description: Retrieves a list of courses with optional filtering
      operationId: getCourses
      tags:
        - Courses
      parameters:
        - name: status
          in: query
          description: Filter by course status
          schema:
            type: string
            enum: [published, unpublished, draft]
        - name: featured
          in: query
          description: Filter featured courses
          schema:
            type: boolean
        - name: free
          in: query
          description: Filter free courses
          schema:
            type: boolean
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/CourseList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create a new course
      description: Creates a new course
      operationId: createCourse
      tags:
        - Courses
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                alias:
                  type: string
                shortDescription:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: binary
                featured:
                  type: boolean
                  default: false
                free:
                  type: boolean
                  default: true
                certificateTerm:
                  type: string
                  enum: [NA, PASS_ALL, COMPLETE_ALL]
                certificateId:
                  type: string
                startDatetime:
                  type: string
                  format: date-time
                endDatetime:
                  type: string
                  format: date-time
                adminApproval:
                  type: boolean
                  default: false
                autoEnroll:
                  type: boolean
                  default: false
                status:
                  type: string
                  enum: [draft, published, unpublished]
                  default: draft
                params:
                  type: string
                  description: JSON string of additional parameters
              required:
                - title
                - shortDescription
                - description
                - startDatetime
                - endDatetime
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      responseCode:
                        type: integer
                        example: 201
                      result:
                        $ref: '#/components/schemas/Course'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{courseId}:
    get:
      summary: Get course by ID
      description: Retrieves a specific course by ID
      operationId: getCourseById
      tags:
        - Courses
      parameters:
        - name: courseId
          in: path
          required: true
          description: UUID of the course
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/Course'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Update course
      description: Updates an existing course
      operationId: updateCourse
      tags:
        - Courses
      parameters:
        - name: courseId
          in: path
          required: true
          description: UUID of the course
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                alias:
                  type: string
                shortDescription:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: binary
                featured:
                  type: boolean
                free:
                  type: boolean
                certificateTerm:
                  type: string
                certificateId:
                  type: string
                startDatetime:
                  type: string
                  format: date-time
                endDatetime:
                  type: string
                  format: date-time
                adminApproval:
                  type: boolean
                autoEnroll:
                  type: boolean
                status:
                  type: string
                  enum: [draft, published, unpublished]
                params:
                  type: string
                  description: JSON string of additional parameters
      responses:
        '200':
          description: Course updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/Course'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Delete course
      description: Deletes a course
      operationId: deleteCourse
      tags:
        - Courses
      parameters:
        - name: courseId
          in: path
          required: true
          description: UUID of the course
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Course deleted successfully
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...